requirements.txt
requests
qrcode
python-dotenv
oauthlib
requests-oauthlib
Pillow